requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.2.2
flask==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
importlib-metadata==8.4.0
itsdangerous==2.2.0
jinja2==3.1.4
jiter==0.5.0
MarkupSafe==2.1.5
openai==1.43.0
pydantic==2.8.2
pydantic-core==2.20.1
python-dotenv==1.0.1
requests==2.32.3
sniffio==1.3.1
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.2
werkzeug==3.0.4
zipp==3.20.1
